Moved getDatabase into datbase folder
