+++ app/assets/js/src/database/getImages.ts
@@ -1,4 +1,4 @@
-import { getDatabase } from 'utils/indexedDB';
+import { getDatabase } from './utils';
 import { ObjectStoreName } from './metadata';
 import type { DatabaseData } from './types';
 import { getImagesInternal } from './internal/getImagesInternal';
