+++ app/assets/js/src/database/internal/getDayInternal.test.ts
@@ -6,10 +6,10 @@ import {
 	test,
 } from '@jest/globals';
 
-import { getDatabase } from 'utils/indexedDB';
+import { getDatabase } from '../utils';
+import { ObjectStoreName } from '../metadata';
+import { createTestData } from '../test-utils';
 
-import { ObjectStoreName } from 'database/metadata';
-import { createTestData } from 'database/test-utils';
 import { getDayInternal } from './getDayInternal';
 
 describe('getDayInternal', () => {
