+++ app/assets/js/src/database/internal/getDayTasksForDayInternal.test.ts
@@ -5,9 +5,9 @@ import {
 	test,
 } from '@jest/globals';
 
-import { getDatabase } from 'utils/indexedDB';
-import { createTestData } from 'database/test-utils';
-import { ObjectStoreName } from 'database/metadata';
+import { getDatabase } from '../utils';
+import { createTestData } from '../test-utils';
+import { ObjectStoreName } from '../metadata';
 
 import { getDayTasksForDayInternal } from './getDayTasksForDayInternal';
