+++ app/assets/js/src/database/internal/getDayTasksInternal.test.ts
@@ -4,9 +4,11 @@ import {
 	expect,
 	test,
 } from '@jest/globals';
-import { ObjectStoreName } from 'database/metadata';
-import { createTestData } from 'database/test-utils';
-import { getDatabase } from 'utils/indexedDB';
+
+import { getDatabase } from '../utils';
+import { ObjectStoreName } from '../metadata';
+import { createTestData } from '../test-utils';
+
 import { getDayTasksInternal } from './getDayTasksInternal';
 
 describe('getDayTasksInternal', () => {
