+++ app/assets/js/src/database/internal/getImageByHashInternal.test.ts
@@ -6,10 +6,9 @@ import {
 	test,
 } from '@jest/globals';
 
-import { getDatabase } from 'utils/indexedDB';
-
-import { ObjectStoreName } from 'database/metadata';
-import { createTestData } from 'database/test-utils';
+import { getDatabase } from '../utils';
+import { ObjectStoreName } from '../metadata';
+import { createTestData } from '../test-utils';
 
 import { getImageByHashInternal } from './getImageByHashInternal';
