+++ app/assets/js/src/database/test-utils/createTestData.ts
@@ -1,9 +1,10 @@
 import { clearDatabase } from './clearDatabase';
 
-import { getDatabase, getIdbRequestPromise } from 'utils/indexedDB';
+import { getIdbRequestPromise } from 'utils/indexedDB';
 
-import type { DatabaseData } from 'database/types';
-import { ObjectStoreName } from 'database/metadata';
+import { getDatabase } from '../utils';
+import type { DatabaseData } from '../types';
+import { ObjectStoreName } from '../metadata';
 
 /**
  * **Important!** For use within tests only.
