+++ app/assets/js/src/utils/indexedDB/index.ts
@@ -1,5 +1,4 @@
 export { doDatabaseTransaction } from './doDatabaseTransaction';
-export { getDatabase } from './getDatabase';
 export { getDbDump } from './getDbDump';
 export { getEntries } from './getEntries';
 export { getIdbRequestPromise } from './getIdbRequestPromise';
